+++ test/pathBuilder.test.js
@@ -32,6 +32,12 @@ describe('javascript paths', () => {
 
 });
 
+describe('`logger` path', () => {
+    it('`jsErrorLoggerSubDir` path should be correct', () => {
+        expect(pathBuilder.jsErrorLoggerDir).toBe('./js/shared');
+    });
+});
+
 describe('image paths', () => {
 
     it('src directory path should be correct', () => {